Create an agnostic service - Move useCrossmintModal logics out
