+++ packages/core/base/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./utils/index";
+export * from "./models/types";
+export * from "./services/useCrossmintModalService";
