+++ packages/ui/react-ui/package.json
@@ -33,6 +33,7 @@
         "test": "jest"
     },
     "dependencies": {
+        "@crossmint/client-sdk-base": "0.0.2-alpha.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
         "react-jss": "^10.9.0",
